Fixed the issue with the wrongly assigned tag
Previously the tag of each page was a tuple and not a string, that was causing problems when rendering the pages, or navigating to a page based on tag. The issue was, because behind the assignment of tag in BasePage, there was a single comma ',' so 'tag = tag,'. Now it is removed and works as expected. Additionally changed the way pages are rendered, now rendering is based on dpg.item_show and item_hide and pages are built once, at program start