Finished the implementation of page manegement and applied some fixes
Finished the page management system. Now child windows are properly displayed within the parent window. Changed the way pages are shown, as previously the page build and switch was causing bugs with 'Alias already exists', thrown from the dearpygui library. Currently, after the page is build, a property called is_built (that is declared in the BasePage class, is switched to true. Each time a page switch is triggered, the page manager checks if the page is already built and if it is, then it simply calls the show method of each page class, otherwise build the page first, and then shows it. This ensures a page is built only once.