Implemented authenticator for the goolge APIs
Implemented a function, that is being called on application start, to authenticate for Google APIs. The function will open the browser and prompt the user to authenticate, before proceeding, that is in case, the user hasn't been authenticated before, and no token is generated. If token.json file exits, it gets the token from there.