Continued the implementation of the google_sheets_service and implemented a new higher level service
Implemented new help methods in the google_sheets_service - a method for composing range and a method for dynamic fetching of the last row of the sheet, the methods are helpful for dynamic operations. Additionally, implemented a service to handle all tasks related to the job tracking, the service utilises the google_sheets_service. The implementation of the service, was mostly due to scalability concerns, the service allows for more modulised, not so coupled application structure, allowing for changes on all levels. The new services also allows for more dynamic implementaion.