Debugged and fixed the problems with authenticating for google and fetching data.
Fixed the problem with authenticating for google, the main problem, causing all issues, was related to how I compose the credentials object, I was passing the dictionary to the build method, when it was expecting a Credentials object. Now that is fixed. Additionally fixed a problem with fetching data, there were some async functions that were not awaited properly